CSV to list conversion working, started to refractor

+++ src/Row.java
@@ -1,17 +1,17 @@
 public class Row {
     public String Recipe;
-    public int ProductionTime;
+    public double ProductionTime;
     public String Machine;
     public String Output1;
-    public int RatePerMinOut1;
+    public double RatePerMinOut1;
     public String Output2;
-    public int RatePerMinOut2;
+    public double RatePerMinOut2;
     public String Input1;
-    public int RateIn1;
+    public double RateIn1;
     public String Input2;
-    public int RateIn2;
+    public double RateIn2;
     public String Input3;
-    public int RateIn3;
+    public double RateIn3;
     public String Input4;
-    public int RateIn4;
+    public double RateIn4;
 }
